docs: add ways-to-contribute.md (#26)
## Description

Add ways to contribute documentation. 
This is meant to be an initial version, I am not looking for a perfect document at the moment.

Fixes #25 

## Why

This will be helpful to answer a common question from the community, on how to contribute 🎉 
We can link this document to newcomers, so they know all the types of contributions we welcome and help them get started faster!